fix: remove cloudpickle (#56)
When cloudpickle is in use, multiprocessing fails to pass object
(e.g., pipe or queue objects) as child process's argument correctly
and causes OSError - bad file descriptor error. We used cloudpickle
to serialize instance's method and use the instance's method (e.g.,
w.run as Process's first function), but we can have a workaround for that.
So, we remove it from the source code.