refactor: multiple jobs within the agent (#40)
Refactor was needed for the agent to be able to run multiple jobs. Now, whenever a new config is being received by the agent, it saves job config with the job_id and starts the job. Also, job update is possible, if the job_id exists within the agent.
Also, added the ability to run the same job again, without restarting the agent