refactor: simplify protobuf
fetch and get_job_action in the protobuf have has similar functionality.
But they are separately defined, which requires maintaining more lines of
redundant code. Now fetch and get_job_action are merged into fetch. And,
the necessary modifications in controller and agent is made accordingly.
Now the code is easy to follow.