feat: online job reconfiguration
Job can be reconfigured in an online fashion. For instance, during the runtime,
a job for resnet152/linear.yaml configuration can be reconfigured to support
resnet152/diamond.yaml configuration, vice versa.

The caveat with the current version is that the configurations used must be
compatible for reconfiguration. Otherwise, the online reconfiguration leads
to an undetermined behavior; it most likely to make the job fail.
Currently, resnet152's linear and diamond configurations are modified to be
compatible for the online reconfiguration.

Also, the logic for computing workers that need to be updated was incorrectly
implemented. That is fixed in this commit.

Finally, the current job state management in the controller is incomplete; e.g.,
it doesn't update job's state to "started" after the job started. But its state
is kept as "starting". There is no logic to make transition from "starting" to
"started". Therefore, the update job action raises an error because the update
job action is not allowed under "starting" state. The job state management
logic requires substantial refactoring. So, to enable the online reconfiguration
feature under the current implementation, the update job action is added to
"starting" state as an allowed action, which is a temporary solution. This will
be handled properly when the job state managment logic is completely revamped.