refactor: use agent context to send GRPC messages
In order to make controller more dynamic in terms of multiple agents, we had to rely on agent context. Therefore, whenever a new config is received via CLI, a GRPC message is being sent to the available agent using context writer. Also, the connection with the agent is kept alive using context's event.