feat: dynamic port support
When a config is being received, the controller sends a request to the agent in order to get available port numbers for workers.
Agent returns to the controller the available port numbers with length of double the number of workers, due to channel connection that is needed.
When the controller receives the port numbers, it updates the config file and sends it back to the agent to start or update the job.
For this flow, JobState class was refactored to store the incoming config.