fix: invalid function call (#77)
cleanup_tx_q function's name in router.py was incorrectly specified.
It caused a silient failure since workers were invoked via multiprocessing.
In this case, an exception in the child process was never displayed on a
terminal. So, this error was not caught. The correct name for the function
is _cleanup_tx_q. This issue is fixed in this commit.