feat: added updated check
When a job is running and it gets a new config, it transitions to updating state.
Since _run_server and _run_worker were called whenever a new config is being received, status messages were inconsistent.
For that, we've moved the status message sending into the actual send and receive calls.
With this, we avoided workers being stuck into a wrong state. e.g. - s-0 was running, with new config s-0 peer was updated, it went back to starting.
Also, logic to check if all workers are running is reused to transition from updating to running.