refactor: update context methods signature
In order to ease the state classes methods handling, we set the request as a local variable in the JobContext every time a new request is being received.
Since sub-state classes have the JobContext class instance as a member variable, they can use the local request variable in their logic.
With this, the entire class / sub-class methods signature synchronization is handled gracefully.