feat: added job status message
Since agent has the information about workers status, it will handle the job status based on the worker status message that it receives.
A new rpc message type was defined to pass the job status to the controller.
Getting the job status and controller handling of the received job status will be handled in other PRs.