feat: implemented agent resources collect
Created CPUStats class to gather CPU related information.
Agent asks for CPU and GPU resources whenever it needs this information.
A new rpc message was created to handle agent's resources.
Also, a new JobAction was created for controller to ask the agent for its resources
Controller doesn't use the agent's resources yet, until we decide how to ask for these.