feat: support auto config attribute in config
The purpose of auto_config flag is to allow the system or not to take care of worker's device and backend attributes. If is set to False, then the device and backend attributes for each worker are mandatory
If set to True, the system will assign to each worker the proper device and backend, based on available agents resources that will deploy these workers.