fix: fixed various issues and updated code to handle agent resources
We added job_id as rpc param for ResourceStats in order for the job_context to know when gathering agent resources for that specific job was finished.
For that, we also added agent_res_event in job_context for the job to know when to proceed with sending config to agents based on resources split.
Therefore, controller will call set_agent_res_processed from job_context for each agent that had resources processed to update their resources_fetched flag.
Fix the type of AgentResources for GPU and VRAM stats.
Created proto_to_stats method in cpu.py.
Called proto_to_stats in controller for each agent resource so we can save the class instance instead of string.
Fixed GPU instance check of RepeatedCompositeContainer from isinstance to type. Imported it from google.protobuf.pyext._message
Also, while debugging the rpc messages with the responses that were received from agents, we identified an issue with fields with default values (0 for int and false for bool attributes), were not included in the RPC message.
A solution was to set them as optional in the message definition so the default values are included in the message. See https://protobuf.dev/programming-guides/proto3/#default.