refactor: update complete job logic for multiple agents (#109)
Added logic to support job completion for workers deployed in multiple agents.
Logic was updated in such way that now every worker should have the status of DONE in order to decide that the job is completed.
Adjusted the code to cover these updates like: handling worker status communication, handle worker status update and cleanup.
Since before deciding that a job is completed some logic needs to be executed, a new state class was added to take care of that.