refactor: added WMA calculus to agent resources
In order to have an average of the agent resources, we need to calculate WMA for these resources. For this, we included cpu and dram stats into the report communication between agent and controller.
Updates in controller for handling cpu and dram resources.
Implemented WMA calculus for numeric agent resources.