feat: static policy
A static deployment policy is implemented. When the policy is enabled,
the controller expects that the configuration has ip address of an agent
in which a worker is going to be deployed. If such information is missing
or doesn't match the actual ip address of the agent, an exception is raised.
An example configuration is also written in this commit. As a part of
implementing static policy, failed state management is also updated.