fix: send resources when agent registers
The agent resources gathering were out of sync, since we set and clear the monitor event in the same method. Now, after wait is done, we clear the event so it will wait again. The issue was that agent was waiting for either CPU or GPU resources for another 10 seconds because the event was prematurely cleared.