feat: added fields validation when auto_config is set to False (#131)
If auto_config is set to False, that means that device and backend attributes needs to be provided. Therefore, we added validation in job config class to check the value and auto_config and if the attributes aren't omitted.
Added error handling on start and update commands to catch the invalid job config.
Also, moved exceptions file one level up since it's used by cmd as well.
An example with auto_config set to True was created without any details about devices and backend options.