feat: decide backend in policy
Decide the backend used by the system based on auto config flag. If the flag is false, the backend used is whatever is set on the worker.
If the flag is true, the system will decide the backend based on the device.
Also, AssignmentData class was updated to keep record of the world info value for it to be used later when patching job config for all agents.
All policy classes were updated with this logic.