feat: config validation with exception handler (#145)
Added backend validation based on device that is specified in the config. For cuda device, nccl and gloo backend are accepted, but for cpu device, only gloo is accepted. Any other backend specified in the config will raise an exception. Handled both post and put API methods.
Added InfscaleException handler so the validation error messages can be forwarded to the client.
Refactored config.py file to split the data conversion from data validation.
Updated conftest file with the required elements.