refactor: packing policy adjustments (#148)
Refactored the way packing policy deploys the workers. We need to select an agent with greatest amount of resources and deploy as many workers as we can on that agent.
Fixed a bug regarding getting server ids. Not all agents that are available in the job context will have workers deployed, therefore we need to check the running agents for the server workers.