feat: worker termination revamp
There are two scenarios when a worker needs to be terminated:
1. when stop job is called, we should terminate workers as soon as we get the command
2. when update job is called and we need to gracefully terminate workers

Added a new message type to differentiate between these two categories.
Update agents' code to send the correct message type, based on the action.

For second scenario, a mechanism to count all requests that are received and sent is implemented. When terminate command is being received, then pipeline is notified about the termination of its worker. The terminated worker needs to wait until the requests count gets to 0, this means that all pending requests were processed.