feat: receiver updates
Whenever an updated config is received and there's a need to terminate a worker, the receiver of the terminate worker should keep accepting requests from it before removing connections.
The receiver will do the cleanup whenever _recv task exists its loop - error raised, this will mean that the sender - terminated worker - was killed.