misc: controlling metrics collection location (#170)
Performance metrics are basically collected in router. However,
To accurately measure the metrics for the server (dispatcher), we
need to start to measure at server's send and recv task instead of
router's _send and _recv. To avoid too much logic change in router,
metrics collector exposes a mechanism to enable/disable collection
in router. Based on the mechanism, server's metrics collection
location is changed.