feat: start with integration tests
Created static test configuration for running a job config using multiple agents in multiple machines. Added ansible YML files to start controller and agents, to start a job and to do the clean-up of the running processes. Created inventory file to specify the hosts that ansible will use to run ssh commands. Also created a script that will run these commands and to orchestrate the tests. Future implementation to make the tests dynamic will be implemented.