refactor: job context restructuring
The current job context relies on config stored in agent data.
This means that each agent has a copy of config for a job.
This can cause unexpected behaviors when a job is updated
(e.g., due to autoscaling). For instance, if a new agent is
assigned for the job update, it doesn't have a current config.
The agent's meta data treats all the worlds in the config as new
worlds to be set up. However some worlds were already configured.
Therefore, for those worlds, there is no new port allocated.
Meanwhile, the new agent tries to allocate new ports for those
worlds, which runs out the port allocated for new worlds. This
leads to an exception.

To prevent future issues, job context is restructured. The job
context only keeps one copy of job config and agents' meta data
doesn't maintain its own copy. This way, one current config and
one new config are used across agents' meta data. At the time of
sending a config to each agent, we create a copy and update
agent-specific fields.