fix: seamless update
When a worker is removed due to job update, other workers connected
to it need to properly update their state. For example, they need to
clean up connections to the worker. In the current implementation,
A sender of the removed worker sometimes fails to finish tensor
transmission and to push the tensor back into an appropriate queue
so that the tensor can be sent again. The sender terminates a tx asyncio
task prematurely. By introducing an synchronization mechanism using
asyncio's event, the issue is now resolved.