feat: dataset replay
Currently, there is no replay of a dataset, meaning that dataset is once
consumed, inference will finish. For long running experiments, we implement
a functionality to replay a dataset. An attribute (replay) is introduced
and set as an integer. Any negative number means that dataset replay is
infinite while a positive integer means the number of replays.