feat: delay for polling job status
In order to properly sync job start and job status endpoint call, we added a delay attribute to the config file. The default value of it is 0. We also conditionally append the delay block into ansible yaml file only if delay attribute is truthy. Added gather_facts to false for all config files to improve the speed of tests since we don't need yet any details about the system yet.