misc: calibration on request generation and metrics
The current implementation computes the request generation rate and
metrics calculation based on batch. In other words, generation rate
is on a per-batch basis although generation rate given as a config
parameter is request rate. So, this is misleading. Also, metrics
calculation is also done on a per-batch basis; for example, no. of
batchs / second as throughput , and no. of batches in woker as queue
level. This can lead to confusing interpretation since the batch
size can be different. Hence, in this commit, we eliminate this
misinterpretation with request-oriented implementation.