feat: check errors and expose them
Added mechanism to properly check for errors while running infscale commands. A temporary test log folder is created. A new ansible task was created to check the logs from this folder for any error that might occur. Cleanup task was updated to remove temporary logs folder. Also the condition for job status was updated because we want to exit when the job is failed. Updated shell command template to redirect the logs to the newly created folder.