feat: worker failure
By deciding wether a worker failed or not, we listen on signals on the MP pipe. Whenever the pipe is closed, the worker will have the FAILED status. The controller receives this status and based on the flow_graph logic, if the job is failed, then remaining workers will be stopped.