refactor: job status (#233)
Refactored the code to rely on received worker status message, instead of agent deciding the job status. Whenever a worker status is received, based on the value, different methods are called, from the current state class instance. Placeholder methods for cond_running and cond_stopped were added to properly handle received worker status and state transitions. A new worker status is introduced for when server is finished and we use FINISHED to notify the controller which will notify all agents about job completion. Also, added TODOs for the code that needs to be removed after this PR is merged.

Co-authored-by: ragaia@cisco.com <ragaia@cisco.com>