chore: remove job state from agent
Removed code related to agent sending the job status to the controller. Whenever DONE, FAILED or TERMINATED worker status is received, worker_manager will remove the data for that user. Also, when all the workers for the given job id are removed from the worker_manager, job_manager does the cleanup.