feat: recovery state
Introduced Recovery job state. Recovery state will be used to recover failed workers. Running state will have cond_recovery and since we want to keep the existing functionality of Running state for when a worker fails, a TOOD was added in cond_recovery which will be removed once the recovery process is implemented. With this, the job will transition to Recovery and it will automatically call cond_failing from the Recovery class to check the job health.