feat: worker recovery on controller
In order to properly recover a failed worker, a continuous resources check is added. We need to keep trying to find resources until the worker can be deployed. After the job transitions to Recovery state, we immediately start asyncio tasks for each failed worker to find agent with available GPU. This task will run until an available GPU is found. After a map between failed worker and agent is created, the current config is updated and it will follow the update job logic. Also, on_exit method was added to base state class which will be called every time we transition to a new state. RecoveryState has this method implemented which will cancel the recover task. Next, we'll need to properly handle recovered worker start on the agent.