feat: controller worker recovery
Added logic to detect recoverable worlds, marked with recover flag. Based on this flag, the re-configuration of the config is complete, meaning that new ports will be assigned to the recovered world info. Logic to reset recover flags from the config was also added. Also, a new command was introduced to notify all agents about a worker failure, so other workers can do the cleanup based on failed worker id.