fix: resources update for update action
Agent's resources were only computed when the start job action was executed. If a job is running and another agent joins at a later time, it's resources weren't recognized by the job context. This issue was raising an insufficient resources error when using autoscaler or when doing a manual job update. A fix was added to resolve this issue.