feat: worker recovery in execution (#252)
Whenever a new serve config is being received, we need to to the cleanup of the existing worlds, based on the serve spec's recover flag and based on world's info recover flag. Logic is added to fetch the list of worlds that need to be recovered and do the cleanup on these worlds.