feat: handle multiple worker failures in recovery
Whenever the job is in RecoveryState, if a subsequent worker failure happens, we need to send a stop command to all workers to be killed and transition the job to failed state since we don't have support for multiple worker failures. Also added TODO to properly handle multiple worker failures.