fix: server deadlock on indirect worker failure
Triggering mechanism that relies on config update is buggy because server is not connected to broken worker directly and it will not have any updated config.
With this PR we rely on check look command received from the controller.
Whenever check look command is received with an empty set of failed worker ids, the job is fully recovered and we can reset inflight requests count and tx event so that the server can continue to send requests.