feat: worker error handler
Created worker error handler class with the primary goal to properly handle silent worker failure. The principle is to have a singleton class instance that will be used in every worker subclass that might need proper error handling in coroutines, threads and so on. Currently the logic is implemented in such way that if an error message is added in the queue, we simply kill the worker process. The usages of this class will be added in another PR, in key points of worker lifecycle.