fix: worker cleanup (#271)
There was a key error while stopping a job due to the fact that the cleanup in worker manager was made in two different places. Once in _handle_worker method that's being called when status_q has a new entry, and once just before populating status_q with a new entry, in WorkerManager. We keep the deletion of worker from the dict in WorkerManager, and only have the job manager cleanup in the agent. With this fix, whenever a valid status message is received from workers in WorkerManager, we do the cleanup, then the agent will do the cleanup in job manager if there's no worker left for that job ID.