fix: worker cleanup
In order to fix the timing issue between worker_manager and agent, we had to refactor the code and to properly establish roles of agent and worker_manager in this worker status update.
After the refactor, worker manager has the role of publisher and state keeper, agent has the role of orchestrator meaning that it will decide when a worker needs to be removed from the state.