feat: remove pipeline
We need a way to remove a pipeline from the config. For that, we created a static method in job config class which will receive a job config and a set of worker ids as arguments. Based on the worker ids, it identifies complete pipelines connecting to server and will remove the affected workers and cleanup remaining worlds from other workers.