feat: pipeline cleanup when no resources
We need to limit recovery attempts whenever we want to recover a job. There's no point on retrying forever to find resources on agents. For that, we implemented an exponential backoff retry mechanism. If there are no resources to recover the workers, we cleanup the config by removing failed pipelines and check if the config is runnable or not. If not, we transition to failed and send a stop command to all remaining workers.