feat: multi rate generator
In order to test scale_in properly, we need a way to dynamically increase or decrease sending rate. For that, we added multi rate generator schedule. The key is the replay on which the rate update should happen and the value is the new rate. Added controller configuration as example on how this can be used.