fix: dataset replay last batch
When using request generator with less sending rate, the server send and server recv were out of sync. With previous implementation, dataset would return None when there were no batches. For that reason, due to longer wait time given by the slower rate server finished processing last batches before getting the info that these are done. The fix includes early detection of the last batch and send that info with the batches, so the server would know which batch is the last one.