refactor: rename recover count attribute
Since we will expand the world duplication count in a next PR, we need to rename recover_count to duplicate_count. Besides the worker failure when we need to increment this count, we need to also update that count for multiple scale_out / in / out scenarios. Second scale_out generated config will have the same world names as the first scale_out, therefore, there will be a world name conflict in multiworld.